feat: implement functions for modules
